# docs/api-coverage/openapi-non-prod.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
...
  "paths" : {
...
    "/api/agent/runs/{runUuid}" : {
      "get" : {
        "operationId" : "getRun_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "runUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAgentRunVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "agent-controller" ]
      }
    },
...
    "/api/v1/workflow-runs/{workflowRunUuid}" : {
      "get" : {
        "operationId" : "getRun",
        "parameters" : [ {
          "in" : "path",
          "name" : "workflowRunUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseWorkflowRunDetailVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "workflow-run-query-controller" ]
      }
    },
    "/api/v1/workflow-runs/{workflowRunUuid}/artifacts" : {
      "get" : {
        "operationId" : "getArtifacts",
        "parameters" : [ {
          "in" : "path",
          "name" : "workflowRunUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListArtifactSummaryVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "workflow-run-query-controller" ]
      }
    },
    "/api/v1/workflow-runs/{workflowRunUuid}/cancel" : {
      "post" : {
        "operationId" : "cancel",
        "parameters" : [ {
          "in" : "path",
          "name" : "workflowRunUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseWorkflowRunCommandVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "workflow-run-command-controller" ]
      }
    },
...
    "/api/v1/workflow-runs/{workflowRunUuid}/retry" : {
      "post" : {
        "operationId" : "retry",
        "parameters" : [ {
          "in" : "path",
          "name" : "workflowRunUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseWorkflowRunCommandVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "workflow-run-command-controller" ]
      }
    },
    "/api/v1/workflow-runs/{workflowRunUuid}/steps" : {
      "get" : {
        "operationId" : "getSteps",
        "parameters" : [ {
          "in" : "path",
          "name" : "workflowRunUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListWorkflowStepReadVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "workflow-run-query-controller" ]
      }
    },
...
    "/api/workflow/game-design/run" : {
      "post" : {
        "operationId" : "run",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WorkflowRunRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseWorkflowRunVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "workflow-controller" ]
      }
    },
    "/api/workflow/{workflowRunUuid}" : {
      "get" : {
        "operationId" : "getWorkflowRun",
        "parameters" : [ {
          "in" : "path",
          "name" : "workflowRunUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseWorkflowRunVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "workflow-controller" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "AgentRunVO" : {
        "properties" : {
          "agentType" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "errorCategory" : {
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "mockState" : {
            "type" : "string"
          },
          "modelName" : {
            "type" : "string"
          },
          "projectId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "projectUuid" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "runUuid" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "timeTakenMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "traceId" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseAgentRunVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/AgentRunVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseListArtifactSummaryVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/ArtifactSummaryVO"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseListWorkflowStepReadVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/WorkflowStepReadVO"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseWorkflowRunCommandVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/WorkflowRunCommandVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseWorkflowRunDetailVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/WorkflowRunDetailVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseWorkflowRunVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/WorkflowRunVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ArtifactSummaryVO" : {
        "properties" : {
          "artifactUuid" : {
            "type" : "string"
          },
          "contentDigest" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "runtimeCapabilityVersion" : {
            "type" : "string"
          },
          "runtimeEligible" : {
            "type" : "boolean"
          },
          "schemaKey" : {
            "type" : "string"
          },
          "schemaVersion" : {
            "type" : "string"
          },
          "sourceArtifactUuid" : {
            "type" : "string"
          },
          "sourceAttempt" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          },
          "stepKey" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "validationSummary" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ErrorSummaryVO" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "WorkflowRunCommandVO" : {
        "properties" : {
          "attempt" : {
            "format" : "int32",
            "type" : "integer"
          },
          "reused" : {
            "type" : "boolean"
          },
          "status" : {
            "type" : "string"
          },
          "workflowRunUuid" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "WorkflowRunDetailVO" : {
        "properties" : {
          "allowedActions" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "artifacts" : {
            "items" : {
              "$ref" : "#/components/schemas/ArtifactSummaryVO"
            },
            "type" : "array"
          },
          "attempt" : {
            "format" : "int32",
            "type" : "integer"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "definitionVersionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorSummaryVO"
          },
          "failedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "lastSequence" : {
            "format" : "int64",
            "type" : "integer"
          },
          "schemaVersion" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "statusVersion" : {
            "format" : "int64",
            "type" : "integer"
          },
          "steps" : {
            "items" : {
              "$ref" : "#/components/schemas/WorkflowStepReadVO"
            },
            "type" : "array"
          },
          "timeTakenMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "workflowRunUuid" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "WorkflowRunRequest" : {
        "properties" : {
          "context" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "idea" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "projectUuid" : {
            "minLength" : 1,
            "type" : "string"
          },
          "title" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "idea", "projectUuid", "title" ],
        "type" : "object"
      },
...
      "WorkflowRunVO" : {
        "properties" : {
          "attempt" : {
            "format" : "int32",
            "type" : "integer"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "inputContent" : {
            "type" : "string"
          },
          "projectId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "projectUuid" : {
            "type" : "string"
          },
          "schemaVersion" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "statusVersion" : {
            "format" : "int64",
            "type" : "integer"
          },
          "steps" : {
            "items" : {
              "$ref" : "#/components/schemas/WorkflowStepVO"
            },
            "type" : "array"
          },
          "summary" : {
            "type" : "string"
          },
          "timeTakenMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "workflowDefinitionVersionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "workflowRunUuid" : {
            "type" : "string"
          },
          "workflowType" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "WorkflowStepReadVO" : {
        "properties" : {
          "agentType" : {
            "type" : "string"
          },
          "artifactType" : {
            "type" : "string"
          },
          "attempt" : {
            "format" : "int32",
            "type" : "integer"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorSummaryVO"
          },
          "finishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "schemaKey" : {
            "type" : "string"
          },
          "schemaVersion" : {
            "type" : "string"
          },
          "startedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "stepKey" : {
            "type" : "string"
          },
          "stepOrder" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timeTakenMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "WorkflowStepVO" : {
        "properties" : {
          "agentRunUuid" : {
            "type" : "string"
          },
          "agentType" : {
            "type" : "string"
          },
          "artifactType" : {
            "type" : "string"
          },
          "artifactUuid" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "stepOrder" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  }
}